requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
distro==1.9.0
elevenlabs==1.13.5
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
Mako==1.3.8
MarkupSafe==3.0.2
microsoft-kiota-abstractions==1.6.6
microsoft-kiota-authentication-azure==1.6.6
microsoft-kiota-http==1.6.6
microsoft-kiota-serialization-form==1.6.6
microsoft-kiota-serialization-json==1.6.6
microsoft-kiota-serialization-multipart==1.6.6
microsoft-kiota-serialization-text==1.6.6
msal==1.31.1
msal-extensions==1.2.0
msgraph-core==1.1.7
msgraph-sdk==1.14.0
multidict==6.1.0
openai==1.57.2
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
pendulum==3.0.0
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
std-uritemplate==2.0.1
time-machine==2.16.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
